@@ -3,14 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.1.3",
-    "@material-ui/icons": "^4.2.1",
+    "@material-ui/core": "^4.9.0",
+    "@material-ui/icons": "^4.5.1",
     "classnames": "^2.2.6",
+    "create-react-app": "^4.0.2",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-scripts": "3.0.1",
-    "react-timeago": "^4.4.0"
+    "react": "17.0.1",
+    "react-dom": "^17.0.1",
+    "react-scripts": "4.0.2",
+    "react-timeago": "^5.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
